Move state of case definition values to definitionview
