+++ frontend/src/Views/CaseView.tsx
@@ -1,7 +1,5 @@
-/* eslint-disable max-len */
 /* eslint-disable camelcase */
 import {
-    Switch,
     Button,
     Icon,
     Menu,
@@ -9,7 +7,6 @@ import {
     Typography,
 } from "@equinor/eds-core-react"
 import {
-    MouseEventHandler,
     useEffect,
     useState,
 } from "react"
@@ -22,20 +19,9 @@ import { Project } from "../models/Project"
 import { Case } from "../models/case/Case"
 import { GetProjectService } from "../Services/ProjectService"
 import CaseAsset from "../Components/Case/CaseAsset"
-import CaseDescription from "../Components/Case/CaseDescription"
-import CaseName from "../Components/Case/CaseName"
-import { unwrapCase, unwrapProjectId } from "../Utils/common"
-import CaseDGDate from "../Components/Case/CaseDGDate"
-import CaseArtificialLift from "../Components/Case/CaseArtificialLift"
-import DGEnum from "../models/DGEnum"
-import ProductionStrategyOverview from "../Components/ProductionStrategyOverview"
-import NumberInput from "../Components/NumberInput"
-import { GetCaseService } from "../Services/CaseService"
-import ExcelUpload from "../Components/ExcelUpload"
+import { unwrapProjectId } from "../Utils/common"
 import DefinitionView from "./DefinitionView"
-import ExplorationView from "./ExplorationView"
 import ExplorationViewTab from "./ExplorationViewTab"
-import { Modal } from "../Components/Modal"
 import { EditCaseInputModal } from "./EditCaseInputModal"
 import CaseCessationCostProfile from "../Components/Case/CaseCessationCostProfile"
 
@@ -46,14 +32,6 @@ const CaseViewDiv = styled.div`
     margin: 2rem;
     display: flex;
     flex-direction: column;
-`
-
-const Wrapper = styled.div`
-    display: flex;
-    > *:not(:last-child) {
-        margin-right: 1rem;
-    }
-    flex-direction: row;
 `
 
 const TopWrapper = styled.div`
@@ -96,13 +74,6 @@ function CaseView() {
     const [caseItem, setCase] = useState<Case>()
     const [activeTab, setActiveTab] = useState<number>(0)
     const { fusionProjectId, caseId } = useParams<Record<string, string | undefined>>()
-    const [artificialLift, setArtificialLift] = useState<Components.Schemas.ArtificialLift>(0)
-    const [prodStratOverview, setProdStratOverview] = useState<Components.Schemas.ProductionStrategyOverview>(0)
-    const [producerCount, setProducerCount] = useState<number>()
-    const [gasInjectorCount, setGasInjectorCount] = useState<number>()
-    const [waterInjectorCount, setWaterInjectorCount] = useState<number>()
-    const [facilitiesAvailability, setFacilitiesAvailability] = useState<number>()
-    const [isReferenceCase, setIsReferenceCase] = useState<boolean | undefined>()
 
     const [isMenuOpen, setIsMenuOpen] = useState<boolean>(false)
     const [element, setElement] = useState<HTMLButtonElement>()
@@ -126,45 +97,13 @@ function CaseView() {
     useEffect(() => {
         if (project !== undefined) {
             const caseResult = project.cases.find((o) => o.id === caseId)
-            if (caseResult !== undefined) {
-                setArtificialLift(caseResult.artificialLift)
-                setProdStratOverview(caseResult.productionStrategyOverview)
-                setFacilitiesAvailability(caseResult?.facilitiesAvailability)
-                setIsReferenceCase(caseResult?.referenceCase ?? false)
-            }
             setCase(caseResult)
-            setProducerCount(caseResult?.producerCount)
-            setGasInjectorCount(caseResult?.gasInjectorCount)
-            setWaterInjectorCount(caseResult?.waterInjectorCount)
-            setFacilitiesAvailability(caseResult?.facilitiesAvailability)
         }
     }, [project])
-
-    useEffect(() => {
-        (async () => {
-            if (caseItem) {
-                const caseDto = Case.Copy(caseItem)
-                caseDto.producerCount = producerCount
-                caseDto.gasInjectorCount = gasInjectorCount
-                caseDto.waterInjectorCount = waterInjectorCount
-                caseDto.facilitiesAvailability = facilitiesAvailability
-                caseDto.referenceCase = isReferenceCase ?? false
-
-                const newProject = await (await GetCaseService()).updateCase(caseDto)
-                setCase(newProject.cases.find((o) => o.id === caseItem.id))
-            }
-        })()
-    }, [producerCount, gasInjectorCount, waterInjectorCount, facilitiesAvailability, isReferenceCase])
 
     const onMoreClick = (target: any) => {
         setElement(target)
         setIsMenuOpen(!isMenuOpen)
-    }
-
-    const switchReference: MouseEventHandler<HTMLInputElement> = () => {
-        if (!isReferenceCase || isReferenceCase === undefined) {
-            setIsReferenceCase(true)
-        } else setIsReferenceCase(false)
     }
 
     if (!project) return null
@@ -241,14 +180,6 @@ function CaseView() {
                                 setProject={setProject}
                                 caseItem={caseItem}
                                 setCase={setCase}
-                                artificialLift={artificialLift}
-                                setArtificialLift={setArtificialLift}
-                                productionStrategyOverview={prodStratOverview}
-                                setProductionStrategyOverview={setProdStratOverview}
-                                switchReference={switchReference}
-                                isReferenceCase={isReferenceCase}
-                                facilitiesAvailability={facilitiesAvailability}
-                                setFacilitiesAvailability={setFacilitiesAvailability}
                             />
                         </StyledTabPanel>
                         <StyledTabPanel>
@@ -270,117 +201,6 @@ function CaseView() {
                         </StyledTabPanel>
                     </Panels>
                 </Tabs>
-
-                <CaseName
-                    caseItem={caseItem}
-                    setProject={setProject}
-                    setCase={setCase}
-                />
-                <ExcelUpload setProject={setProject} setCase={setCase} />
-                <CaseDescription
-                    caseItem={caseItem}
-                    setProject={setProject}
-                    setCase={setCase}
-                />
-                <Switch
-                    onClick={switchReference}
-                    label="Reference case"
-                    readOnly
-                    checked={isReferenceCase ?? false}
-                />
-                <Wrapper>
-                    <CaseDGDate
-                        caseItem={caseItem}
-                        setProject={setProject}
-                        setCase={setCase}
-                        dGType={DGEnum.DG0}
-                        dGName="DG0"
-                    />
-                </Wrapper>
-                <Wrapper>
-                    <CaseDGDate
-                        caseItem={caseItem}
-                        setProject={setProject}
-                        setCase={setCase}
-                        dGType={DGEnum.DG1}
-                        dGName="DG1"
-                    />
-                    <CaseDGDate
-                        caseItem={caseItem}
-                        setProject={setProject}
-                        setCase={setCase}
-                        dGType={DGEnum.DG3}
-                        dGName="DG3"
-                    />
-                </Wrapper>
-                <Wrapper style={{ marginBottom: -35 }}>
-                    <CaseDGDate
-                        caseItem={caseItem}
-                        setProject={setProject}
-                        setCase={setCase}
-                        dGType={DGEnum.DG2}
-                        dGName="DG2"
-                    />
-                    <CaseDGDate
-                        caseItem={caseItem}
-                        setProject={setProject}
-                        setCase={setCase}
-                        dGType={DGEnum.DG4}
-                        dGName="DG4"
-                    />
-                </Wrapper>
-                <CaseArtificialLift
-                    currentValue={artificialLift}
-                    setArtificialLift={setArtificialLift}
-                    setProject={setProject}
-                    caseItem={caseItem}
-                />
-                <ProductionStrategyOverview
-                    currentValue={prodStratOverview}
-                    setProductionStrategyOverview={setProdStratOverview}
-                    setProject={setProject}
-                    caseItem={caseItem}
-                />
-                <DividerLine />
-                <Wrapper style={{ marginBottom: -15 }}>
-                    <CaseArtificialLift
-                        currentValue={artificialLift}
-                        setArtificialLift={setArtificialLift}
-                        setProject={setProject}
-                        caseItem={caseItem}
-                    />
-                    <ProductionStrategyOverview
-                        currentValue={prodStratOverview}
-                        setProductionStrategyOverview={setProdStratOverview}
-                        setProject={setProject}
-                        caseItem={caseItem}
-                    />
-                </Wrapper>
-                <DividerLine />
-                <Wrapper style={{ marginBottom: 45 }}>
-                    <NumberInput
-                        setValue={setProducerCount}
-                        value={producerCount ?? 0}
-                        integer
-                        disabled={false}
-                        label="Producer count"
-                    />
-                    <NumberInput
-                        setValue={setGasInjectorCount}
-                        value={gasInjectorCount ?? 0}
-                        integer
-                        disabled={false}
-                        label="Gas injector count"
-                    />
-                    <NumberInput
-                        setValue={setWaterInjectorCount}
-                        value={waterInjectorCount ?? 0}
-                        integer
-                        disabled={false}
-                        label="Water injector count"
-                    />
-                </Wrapper>
-                <DividerLine />
                 <CaseCessationCostProfile
                     dG4Year={caseItem.DG4Date?.getFullYear()}
                     timeSeries={caseItem.cessationCost}
